hostip: Cache negative name resolves
TODO 1.9: Add failed name resolves to the host cache, so that repeated
connection attempts do not immediately repeat the resolve request. Lifetime
of negative entries is set to half of the positive ones.

Refactor the hostip api so that Curl_resolv_check() also checks the cache,
removing the need for Curl_fetch_addr(). This way hostip callers only need
one function, and don't care if the result came from cache or not.