introducing "curldown" for libcurl man page format
curldown is this new file format for libcurl man pages. It is markdown
inspired with differences:

- Each file has a set of leading headers with meta-data
- Supports a small subset of markdown

cd2nroff is a tool that converts from curldown to standard man
page. This makes us still able to provide and ship regular man pages.

nroff2cd is a tool to convert an existing libcurl man page to
curldown. Run 'make 2cd' in the docs/ dir to get all nroff files
converted to curldown.

To test it:

    cd docs
    make 2cd

Convert all nroff pages to .cd versions.

    ../scripts/cdall.pl libcurl libcurl/opts

Generates .3 versions of all the .cd versions, for the sake of doing
comparisons with the original .3 to verify that the conversion looks
correct.