sasl: use http authentication mechanisms as a default, not a preset
If some are given, sasl maps http authentication options to sasl ones
and merges them to the login options.
This may cause problems with the cli tool that sets the http login
option CURLAUTH_BEARER as a side-effect of --oauth2-bearer, because this
flag maps to more than one sasl mechanisms and the latter cannot be
cleared individually by the login options string.
The solution retained here is to consider http authentication options as
a default (overriding the hardcoded default mask for the protocol) that
is ignored if a login option string is given.

New test 992 checks this.

Fixes #10259