tool: Add option --retry-ignore-server-time
- Add option --retry-ignore-server-time to ignore the server's requested
  retry delay time (aka the optional Retry-After response header).

This option stops the server from overriding curl's retry delay time
before retrying a transient HTTP error (eg Too Many Requests, Gateway
Timeout, etc). It is not polite to ignore the server time however it may
be necessary for some network conditions.

For example, a server may set a retry time of 1 second even though that
may not be practical depending on how busy the server is. The user can
use this option to ignore the server time in favor of --retry-delay time
(or the default algorithmic retry delay time), thereby giving their
transfers a greater chance of success.

Bug: https://curl.se/mail/archive-2024-01/0022.html
Reported-by: Dirk Hünniger

Closes #xxxx