mprintf: fix format prefix I32/I64 for Borland compilers
- Support I32 & I64 (eg: %I64d) for all Win32 builds.

Prior to this change mprintf support for the I format prefix, which is a
Microsoft extension, was dependent on the compiler used.

When Borland support was removed in fd7ef00f the prefix was then no
longer supported for that compiler; however since it's still possible to
build with Borland I'm restoring support for the prefix in this way.

Reported-by: Paweł Witas

Fixes https://github.com/curl/curl/issues/12944
Closes #xxxx