fopen: fix narrowing conversion warning on 32-bit Android
This was fixed in commit 06dc599405fd8a87147eddcdb89dab62ce200d9b, but came back in commit
03cb1ff4d629d6110dab787c75c187626d58323d.

When building for 32-bit ARM or x86 Android, `st_mode` is defined as
`unsigned int` instead of `mode_t`, resulting in a
`-Wimplicit-int-conversion` clang warning because `mode_t` is
`unsigned short`. Add a cast to silence the warning, but only for
Android builds, because other architectures and platforms are not
affected.

Ref: https://android.googlesource.com/platform/bionic/+/refs/tags/ndk-r25c/libc/include/sys/stat.h#86