cmake: fix `CURL_WINDOWS_SSPI=ON` option
Prior to this change `CURL_WINDOWS_SSPI` was accidentally forced
`OFF` when building without the Schannel TLS backend.

This in turn may have caused Kerberos, SPNEGO and SSPI features
disappearing even with `CURL_WINDOWS_SSPI=ON` set.

This patch fixes it by using the `CURL_USE_SCHANNEL` setting as a
default for `CURL_WINDOWS_SSPI`, but allowing a manual override.

Ref: #13056
Closes #xxxxx