m4: reposition USE_RUSTLS="yes" for pkg-config
It's necessary to set this var to "yes" _after_ AC_DEFINE and AC_SUBST
in order for a later `test` to pass so that `check_for_ca_bundle=1`
ends up being set. This is in turn required for the default CA
certificate bundle to be set when building w/ rustls & pkg-config.