tool_cb_hdr: allow --remote-header-name for all http responses
- Use the content-disposition filename (server-specified filename) for
  non-2xx responses.

For example, a server may send a content-disposition filename header
with a redirect reply (3xx) but not with the final response (2xx).
Without this change curl would ignore the server's specified filename
and continue to use the filename extracted from the user-specified URL.

This is a partial revert of 75d79a44, which limited content-disposition
and etag header processing to 2xx response codes. There's no explanation
for why that was done.

Reported-by: Morgan Willcock

Fixes https://github.com/curl/curl/issues/13302
Closes #xxxx