cmake: allow disabling `ENABLE_CURLDEBUG` with `ENABLE_DEBUG=ON`
Before this patch, `ENABLE_CURLDEBUG` (memory tracking) was
unconditionally enabled when `ENABLE_DEBUGBUILD` was set. This made
testing some build configurations complicated. To fix this, after this
patch we only enable `ENABLE_CURLDEBUG` if not already defined by the
user.

This allows to use this config:
ENABLE_DEBUGBUILD=ON
ENABLE_CURLDEBUG=OFF
to enable debug features, without also enabling memory tracking.

This is important because some other build methods allow to set
one of these features but not the other. This patch allows to test
any combination with CMake.

This makes it unnecessary to use the workaround of passing
`-DDEBUGBUILD` via `CMAKE_C_FLAGS`. Which has the disadvantage
that our CMake logic cannot easily detect it, e.g. for disabling
symbol hiding on Windows for `ENABLE_DEBUG`/`DEBUGBUILD` builds.

Cherry-picked from #13718
Closes #xxxxx