aws-sigv4: url encoding on canonical path
I refactor canon_query, so I could use the encoding part
of the function to use it in the path.

As the path doesn't encode '/', but encode '=', I had to
add some conditions to know If I was doing the query or path encoding.

Also, instead of adding a `bool in_path` variable, I use `bool *found_equals`
to know if the function was called for the query or path,
as found_equals is used only in query_encoding.

fix: #13754

Signed-off-by: Matthias Gatto <matthias.gatto@outscale.com>