cd2nroff: use an empty "##" to signal end of .IP sequence
Like when we list a series of options and then want to add "normal" text
again afterwards.

Without this, the indentation level wrongly continues even after the
final "##" header, making following text wrongly appear to belong to the
header above.

Adjusted CURLOPT_SSLVERSION.md to use this to show the idea.

Fixes #13803
Reported-by: Jay Satiro