socketpair: add `eventfd` and use `SOCK_NONBLOCK` for `socketpair()`
Currently, we use `pipe` for `wakeup_create`, which requires ***two***
file descriptors. Furthermore, given its complexity inside, `pipe` is a
bit heavyweight for just a simple event wait/notify mechanism.

`eventfd` would be a more suitable solution for this kind of scenario,
kernel also advocates for developers to use `eventfd` instead of `pipe`
in some simple use cases:
> Applications can use an eventfd file descriptor instead of a pipe
   (see pipe(2) in all cases where a pipe is used simply to signal
    events.  The kernel overhead of an eventfd file descriptor is
    much lower than that of a pipe, and only one file descriptor is
    required (versus the two required for a pipe).

This PR adds the new backend of `eventfd` for `wakeup_create` and uses
it where available, eliminating the overhead of `pipe`. Also, it optimizes
the `wakeup_create` to eliminate the system calls that make file descriptors
non-blocking by moving the logic of setting non-blocking flags on file
descriptors to `socketpair.c` and using `SOCK_NONBLOCK` for `socketpair(2)`,
`EFD_NONBLOCK` for `eventfd(2)`.

Ref:
https://man7.org/linux/man-pages/man7/pipe.7.html
https://man7.org/linux/man-pages/man2/eventfd.2.html
https://man7.org/linux/man-pages/man2/socketpair.2.html
https://www.gnu.org/software/gnulib/manual/html_node/eventfd.html