socket: use SOCK_NONBLOCK to eliminate extra system call
Every time function `cf_socket_open()` is called to create a socket,
`curlx_nonblock()` is called to make that socket non-blocking.
And `curlx_nonblock()` will cost us 1 or 2 system calls (2 for `fcntl()`,
1 for `ioctl()`, etc.), meanwhile, tucking `SOCK_NONBLOCK` and `SOCK_CLOEXEC`
into the `type` argument for `socket()` is widely supported across UNIX-like OS:
Linux, *BSD, Solaris, etc. With that ability, we can save 1 or 2 system calls
on each socket.

Another change in this PR is to eliminate the redundant `curlx_nonblock()` call
on the socket in `cf_udp_setup_quic()` as that socket created by `cf_socket_open()`
is already non-blocking.

Ref:
https://man7.org/linux/man-pages/man2/socket.2.html
https://man.freebsd.org/cgi/man.cgi?socket(2)
https://man.dragonflybsd.org/?command=socket&section=2
https://man.netbsd.org/socket.2
https://man.openbsd.org/socket
https://docs.oracle.com/cd/E88353_01/html/E37843/socket-3c.html
https://illumos.org/man/3SOCKET/socket
...